Added a account existence check for deletion.

+++ routes/admin/accounts.js
@@ -75,8 +75,14 @@ router.post("/:id", auth.session(true), async (req, res) => {
     })
 })
 
-router.delete("/:id", auth.session(true), (req, res) => {
+router.delete("/:id", auth.session(true), async (req, res) => {
     const accountId = req.params.id;
+
+    const editedAccountCheck = await dbQuery("SELECT COUNT(*) AS count FROM accounts WHERE id = ?", [accountId])
+    exists = editedAccountCheck[0].count > 0;
+    if(!exists){
+        return res.status(400).json({ error: "Bad Request", message: "The account you attempted to delete does not exist." })
+    }
 
     db.query("DELETE FROM accounts WHERE id = ?", [accountId], (err, results) => {
         if(err) throw err;
